@@ -1,4 +1,3 @@
-import React from 'react';
 import {userReducer} from './user-reducer';
 
 test('user reducer should increment only age', () => {
@@ -12,18 +11,18 @@ test('user reducer should increment only age', () => {
 
 test('user reducer should increment only childrenCount', () => {
     const startState = { age: 20, childrenCount: 2, name: 'Dimych' };
+
     const endState = userReducer(startState, { type: 'INCREMENT-CHILDREN-COUNT' })
 
-    expect(endState.age).toBe(20);
     expect(endState.childrenCount).toBe(3);
+    expect(endState.age).toBe(20);
 });
 
 test('user reducer should change name of user', () => {
     const startState = { name: 'Dimych', age: 20, childrenCount: 2 };
     const newName = 'Viktor';
+
     const endState = userReducer(startState, { type: 'CHANGE-NAME', newName: newName })
 
     expect(endState.name).toBe(newName);
 });
-
-
