@@ -8,28 +8,24 @@ type ActionType = {
     [key: string]: any
 }
 
-// меня вызовут и дадут мне стейт (почти всегда объект)
-// и инструкцию (action, тоже объект)
-// согласно прописаному type в этом action (инструкции) я поменяю state
-export const userReducer = (state: StateType, action: ActionType) => {
+
+export const userReducer = (state: StateType, action: ActionType): StateType => {
     switch (action.type) {
         case 'INCREMENT-AGE':
-            let newState = {...state};//делаем копию
-            newState.age = state.age + 1;// у копии имеем право менять св-во
-            return newState;//возвращаем копию
+            let newState = {...state};
+            newState.age = state.age + 1;
+            return newState;
         case 'INCREMENT-CHILDREN-COUNT':
-            // а можно без создания переменных промежуточных (делайте, как вам понятнее)
             return {
                 ...state,
                 childrenCount: state.childrenCount + 1
-            };
+            }
         case 'CHANGE-NAME':
-            // а можно без  создания переменных промежуточных (делайте, как вам понятнее)
             return {
                 ...state,
                 name: action.newName
-            };
+            }
         default:
-            throw new Error("I don't understand this type")
+            throw new Error("I don't understand this action type")
     }
 }
